docs(docs): shape the example theme picker like a v0 compound
createExampleTheme spreads the page toggle and overrides writes. The example
root provides that context; the menu consumes it. No cloned store, no
controller prop.
