+++ apps/docs/src/components/docs/DocsExampleThemeMenu.vue
@@ -1,29 +1,17 @@
 <script setup lang="ts">
   // Framework
-  import { Popover } from '@vuetify/v0'
+  import { Button, Popover } from '@vuetify/v0'
 
   // Composables
-  import { provideThemeToggle } from '@/composables/useThemeToggle'
+  import { useThemeToggleController } from '@/composables/useThemeToggle'
 
   // Utilities
-  import { shallowRef, toRef, useId } from 'vue'
+  import { shallowRef, useId } from 'vue'
 
-  // Types
-  import type { UseLocalThemeToggleReturn } from '@/composables/useLocalThemeToggle'
-
-  export interface DocsExampleThemeMenuProps {
-    controller: UseLocalThemeToggleReturn
-  }
-
-  const props = defineProps<DocsExampleThemeMenuProps>()
-
-  provideThemeToggle(props.controller)
+  const toggle = useThemeToggleController()
 
   const uid = useId()
   const open = shallowRef(false)
-  const icon = toRef(() => props.controller.icon.value)
-  const title = toRef(() => props.controller.title.value)
-  const overridden = toRef(() => props.controller.isOverridden.value)
 </script>
 
 <template>
@@ -32,13 +20,14 @@
       as="span"
       class="mt-[8px] me-[8px] inline-flex"
       position-area="bottom"
-      :text="title"
+      :text="toggle.title.value"
     >
       <Popover.Activator
         aria-label="Example theme"
         class="bg-surface-tint text-on-surface-tint pa-1 inline-flex rounded-none rounded-bl-[0.375rem] rounded-tr-[0.375rem] hover:bg-surface-variant transition-all cursor-pointer"
+        :data-state="open ? 'open' : undefined"
       >
-        <AppIcon :icon />
+        <AppIcon :icon="toggle.icon.value" />
       </Popover.Activator>
     </AppTooltip>
 
@@ -51,14 +40,13 @@
       <div class="flex items-center justify-between mb-3 ps-1">
         <span class="text-xs font-semibold text-on-surface">Theme</span>
 
-        <button
-          v-if="overridden"
+        <Button.Root
+          v-if="toggle.isOverridden?.value"
           class="text-xs font-medium text-primary cursor-pointer"
-          type="button"
-          @click="props.controller.reset()"
+          @click="toggle.reset?.()"
         >
           Follow page
-        </button>
+        </Button.Root>
       </div>
 
       <AppThemeMenu />
